+++ app/src/components/index.tsx
@@ -37,4 +37,4 @@ export * from "./textfield";
 export * from "./tag";
 export * from "./overlay";
 export * from "./datetime";
-export * from "./radioGroup";
+export * from "./radio";
